requirements.txt
langchain
langchain_openai
Flask
pytest
python-dotenv